Fix span start times being the same
There was a bug that was incorrectly adding a "milliseconds" offset
to the start time of child spans, when it was expected in microseconds.
Fixed that.

Also, associated the specified tags in topology spec with the span
rather than with the service. If needed, we can create a separate
service tags (and maybe even instance tags) section in the spec.